requirements.txt
aiohttp==3.7.4.post0
alembic==1.4.3
altair==4.1.0
anyio==3.2.1
apache-airflow==2.1.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
appdirs==1.4.4
argcomplete==1.12.1
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
autoviz==0.0.83
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.3.0
blinker==1.4
bokeh==2.3.3
Bottleneck==1.3.2
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
cattrs==1.1.0
certifi==2020.11.8
certipy==0.1.3
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
cloudpickle==1.6.0
colorama==0.4.4
colorcet==2.0.6
colorlog==4.0.2
commonmark==0.9.1
configparse==0.1.5
configparser==3.5.3
connexion==2.7.0
croniter==0.3.36
cryptography==3.4.7
csscompressor==0.9.5
cycler==0.10.0
daiquiri==3.0.0
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-colorscales==0.0.4
dash-core-components==1.16.0
dash-daq==0.5.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
dask==2.30.0
dataclasses==0.6
dataframe-sql==0.4.0
dataprep==0.3.0
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.12
dill==0.3.3
dnspython==2.0.0
docutils==0.16
dtale==1.51.0
email-validator==1.1.2
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==3.3.0
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.10.1
Flask-Compress==1.10.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
flask-ngrok==0.0.25
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
fsspec==2021.6.1
funcsigs==1.0.2
future==0.18.2
gevent==21.1.2
google-auth==1.30.0
google-auth-httplib2==0.1.0
graphviz==0.14.2
greenlet==1.1.0
grpcio==1.37.1
gunicorn==19.10.0
h11==0.12.0
HeapDict==1.0.1
holoviews==1.14.4
htmlmin==0.1.12
httpcore==0.13.6
httplib2==0.19.1
httpx==0.18.2
hvplot==0.7.2
ibis-framework==1.4.0
idna==2.10
ImageHash==4.2.0
importlib-metadata==4.6.0
importlib-resources==1.5.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso3166==1.0.1
iso8601==0.1.13
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.2
joblib==1.0.1
json-merge-patch==0.2
json5==0.9.6
jsonpath-ng==1.5.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-datatables==0.3.9
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-nbutils==0.1.3
jupyter-require==0.6.1
jupyter-server==1.9.0
jupyter-server-proxy==3.1.0
jupyter-telemetry==0.1.0
jupyterhub==1.4.1
jupyterlab==3.0.16
jupyterlab-darkside-ui==0.1.9
jupyterlab-legos-ui==0.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
kaleido==0.2.1
keyring==23.0.1
kiwisolver==1.3.1
lark-parser==0.8.5
lazy-object-proxy==1.4.3
lesscpy==0.14.0
locket==0.2.1
lockfile==0.12.2
lux==0.8.2
lux-api==0.3.1
lux-widget==0.1.6
lxml==4.6.3
lz4==3.1.3
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matlab-kernel==0.16.11
matplotlib==3.3.4
matplotlib-inline==0.1.2
matplotlip==0.2
metakernel==0.27.5
Metaphone==0.6
missingno==0.4.2
mistune==0.8.4
mock==4.0.3
msgpack==1.0.2
multidict==5.1.0
multimethod==1.4
multipledispatch==0.6.0
mypy-extensions==0.4.3
natsort==7.0.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.4
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.7
packaging==20.4
pamela==1.0.0
pandas==1.3.0
pandas-flavor==0.2.0
pandas-profiling==3.0.0
pandocfilters==1.4.3
panel==0.11.3
param==1.11.1
parso==0.7.1
partd==1.2.0
pathspec==0.8.1
patsy==0.5.1
pendulum==2.1.2
pep562==1.0
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.2.0
pivottablejs==0.9.0
pkginfo==1.7.0
plotly==5.1.0
pluggy==0.13.1
ply==3.11
ppscore==1.2.0
prison==0.1.3
probableparsing==0.0.1
prometheus-client==0.11.0
prompt-toolkit==2.0.10
psutil==5.7.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pulsar==2.0.2
pulsar-odm==0.7.0
py==1.9.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyct==0.4.8
pydantic==1.8.2
PyGithub==1.54.1
Pygments==2.7.2
pyjanitor==0.20.14
PyJWT==1.7.1
PyMySQL==0.10.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
python-crfsuite==0.9.7
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==2.0.1
python-Levenshtein==0.12.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
pytzdata==2020.1
pyviz-comms==2.1.0
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.1.0
qgrid==1.3.1
qtconsole==5.1.1
QtPy==1.9.0
readme-renderer==29.0
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==9.2.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
SecretStorage==3.3.1
Send2Trash==1.7.1
setproctitle==1.1.10
simpervisor==0.4
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.1
sql-to-ibis==0.4.0
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
squarify==0.4.3
statsmodels==0.12.2
strsimpy==0.2.0
swagger-ui-bundle==0.0.8
sweetviz==2.1.3
tabulate==0.8.7
tangled-up-in-unicode==0.1.0
tblib==1.7.0
tenacity==6.2.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
thrift==0.13.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
twine==3.4.1
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
usaddress==0.5.10
visions==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
wsgi-oauth2==0.2.2
WTForms==2.3.3
wurlitzer==2.1.0
xarray==0.18.2
xgboost==1.4.2
xlrd==2.0.1
yarl==1.6.3
zict==2.0.0
zipp==3.4.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.interface==5.4.0
